Refactor folder structure and data imports
