--- client/src/components/playground.css
@@ -1,69 +1,0 @@
-/* resets the default margin and padding of the page */
-* {
-    padding: 0;
-    margin: 0;
-  }
-  
-  .about-me {
-    /* display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    background-color: #f9f9f9;
-    min-height: 100vh;
-    font-family: 'Roboto', sans-serif; */
-  }
-  
-  .profile-picture {
-    width: 25%;
-    height: 25%;
-    border-radius: 10% 30%;
-    object-fit: cover;
-    border: 0.5em solid #1098ad;
-  }
-  
-  h2 {
-    font-size: 2em;
-    color: #1098ad;
-    padding: 20px;
-  }
-  
-  .bubble-container {
-    /* position: relative; */
-    display: flex;
-  }
-  
-  .bubble {
-    width: 6em;
-    height: 6em;
-    background-color: #1098ad;
-    color: white;
-    font-weight: bold;
-    font-size: large;
-    padding: 1em;
-    margin: 20px;
-    border-radius: 50%;
-    border: 2px solid #007b83;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    cursor: pointer;
-    /* position: absolute; 
-    top: 30px;
-    left: 50px;*/
-  }
-  
-  .bubble:hover {
-    background-color: #79e8f0;
-    border: 2px solid #42ef65;
-    color: rgb(239, 122, 122);
-  }
-  
-  .blue {
-    /* TODO> adding colors for each of the circles, 
-    and then removing some props from the bubble selector */
-    background-color: #1098ad;
-    border: 2px solid #007b83;
-  }
-
